requirements.txt
fastapi==0.104.0
scrapy==2.11.0
uvicorn==0.23.2
pylint==3.0.2
loguru==0.7.2
starlette==0.27.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
beanie==1.24.0
pandas==2.2.1
python-multipart==0.0.9